+++ lib/helpful_components.dart
@@ -4,7 +4,7 @@ export 'src/pop_up_scope.dart';
 export 'src/boolean_dialog.dart';
 export 'src/common_alert_dialog.dart';
 export 'src/common_snapshot_responses.dart';
-export 'src/custom_pop_up_menu.dart';
+export 'src/pop_up.dart';
 export 'src/future_dialog.dart';
 export 'src/helpful_components_theme.dart';
 export 'src/lazy_builder.dart';